pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "python-sapcommissions"
authors = [{ name = "Niels Perfors", email = "niels.perfors1987@gmail.com" }]
description = "Client for SAP Commissions"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["SAP Commissions", "Incentive Management"]
license = { text = "MIT" }
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.0.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.0",
]
dynamic = ["version"]

[project.scripts]
sap = "sapcommissions.__main__:cli"

[project.urls]
Homepage = "https://github.com/niro1987/python-sapcommissions"
Repository = "https://github.com/niro1987/python-sapcommissions.git"
Issues = "https://github.com/niro1987/python-sapcommissions/issues"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools_scm]

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests"]

[tool.setuptools.package-data]
"sapcommissions" = ["py.typed"]

[tool.pylint.MAIN]
py-version = "3.11"
ignore = ["tests"]
load-plugins = ["pylint.extensions.code_style", "pylint.extensions.typing"]

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "C0103", # invalid-name
]
enable = ["use-symbolic-message-instead"]

[tool.pylint.REPORTS]
score = false

[tool.pylint.FORMAT]
expected-line-ending-format = "LF"

[tool.pylint.TYPING]
runtime-typing = false

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".git"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s.%(msecs)03d %(levelname)-8s %(name)25s:%(lineno)-4s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "tests/test.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s.%(msecs)03d %(levelname)-8s %(name)25s:%(lineno)-4s %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff.lint]
select = [
    "B002",    # Python does not support the unary prefix increment
    "B007",    # Loop control variable {name} not used within loop body
    "B014",    # Exception handler with duplicate exception
    "B023",    # Function definition does not bind loop variable {name}
    "B026",    # Star-arg unpacking after a keyword argument is strongly discouraged
    "C",       # complexity
    "COM818",  # Trailing comma on bare tuple prohibited
    "D",       # docstrings
    "DTZ003",  # Use datetime.now(tz=) instead of datetime.utcnow()
    "DTZ004",  # Use datetime.fromtimestamp(ts, tz=) instead of datetime.utcfromtimestamp(ts)
    "E",       # pycodestyle
    "F",       # pyflakes/autoflake
    "G",       # flake8-logging-format
    "I",       # isort
    "ICN001",  # import concentions; {name} should be imported as {asname}
    "N804",    # First argument of a class method should be named cls
    "N805",    # First argument of a method should be named self
    "S307",    # No builtin eval() allowed
    "PGH004",  # Use specific rule codes when using noqa
    "PLC0414", # Useless import alias. Import alias does not rename original package.
    "PLC",     # pylint
    "PLE",     # pylint
    "PLR",     # pylint
    "PLW",     # pylint
    "Q000",    # Double quotes found but single quotes preferred
    "RUF006",  # Store a reference to the return value of asyncio.create_task
    "S102",    # Use of exec detected
    "S103",    # bad-file-permissions
    "S108",    # hardcoded-temp-file
    "S306",    # suspicious-mktemp-usage
    "S307",    # suspicious-eval-usage
    "S313",    # suspicious-xmlc-element-tree-usage
    "S314",    # suspicious-xml-element-tree-usage
    "S315",    # suspicious-xml-expat-reader-usage
    "S316",    # suspicious-xml-expat-builder-usage
    "S317",    # suspicious-xml-sax-usage
    "S318",    # suspicious-xml-mini-dom-usage
    "S319",    # suspicious-xml-pull-dom-usage
    "S320",    # suspicious-xmle-tree-usage
    "S601",    # paramiko-call
    "S602",    # subprocess-popen-with-shell-equals-true
    "S604",    # call-with-shell-equals-true
    "S608",    # hardcoded-sql-expression
    "S609",    # unix-command-wildcard-injection
    "SIM105",  # Use contextlib.suppress({exception}) instead of try-except-pass
    "SIM117",  # Merge with-statements that use the same scope
    "SIM118",  # Use {key} in {dict} instead of {key} in {dict}.keys()
    "SIM201",  # Use {left} != {right} instead of not {left} == {right}
    "SIM208",  # Use {expr} instead of not (not {expr})
    "SIM212",  # Use {a} if {a} else {b} instead of {b} if not {a} else {a}
    "SIM300",  # Yoda conditions. Use 'age == 42' instead of '42 == age'.
    "SIM401",  # Use get from dict with default instead of an if block
    "T100",    # Trace found: {name} used
    "T20",     # flake8-print
    "TID251",  # Banned imports
    "TRY004",  # Prefer TypeError exception for invalid type
    "B904",    # Use raise from to specify exception cause
    "TRY302",  # Remove exception handler; error is immediately re-raised
    "UP",      # pyupgrade
    "W",       # pycodestyle
]
ignore = [
    "D202", # No blank lines allowed after function docstring
    "D203", # 1 blank line required before class docstring
    "E501", # line too long
    "E731", # do not assign a lambda expression, use a def
    "D213", # Multi-line docstring summary should start at the second line

    # May conflict with the formatter, https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "ISC001",
    "ISC002",
]

[tool.ruff.lint.isort]
known-first-party = ["sapcommissions"]
combine-as-imports = true
split-on-trailing-comma = true
